Add more benchmarks:
- Push/PopLocalFrame overhead benchmark — such operations
are *required* if one attaches a long-running native thread
to the JVM because there is no 'return-from-native-method'
event when created local references are freed, hence no way for
the JVM to know that the local references are no longer used
in the native code.

- A benchmark of the overhead of attaching and detaching
a native thread. It is expectedly HUGE — two orders of magnitude
higher than calling a single Java method using unchecked APIs
(e.g., `jni_call_static_unsafe`).