Detach daemon threads automatically
This is a breaking API change that is required to make the library more safe for use.
Daemon threads are not special in regard to cleanup: they must be detached in the
very same way as common threads (by calling `DetachCurrentThread`). Not managing to do so
may result in unexpected behaviour and memory leaks.