Allow JNIEnv lifetimes to be more flexible (#165)
In some cases, it was only possible to call certain methods if the
lifetime of the JNIEnv borrow was the same of the JNIEnv type (i.e., in
`&'a JNIEnv<'b>`, `'a` had to be equal to `'b`). This PR changes this so
that `'a: 'b` (i.e., the type lives at least as much as the borrow).

This lead to some changes to certain wrapper types (`JList`, `JMap` and
`JavaStr`), which now require both lifetimes to be explicit.

Another requirement was that some implicit lifetimes had to become
explicit, otherwise they would be infered to be the lifetime of `&self`,
i.e., the borrow, not the type.