Use repr(transparent) and not repr(C) for transparent wrappers (#170)
#[repr(C)] struct Foo(T) isn't guaranteed to behave identically to T, 
but #[repr(transparent)] is. In practice, these two are often different,
although for pointers (which is how they're used in jni-rs) they're typically the same.

This changes the uses of #[repr(C)] where identical behavior
is desired/required to use #[repr(transparent)].

In practice this I think this won't matter for most modern systems,
since these are wrappers around pointers (it might matter on arm64
if you have a method with many arguments), and they're only present
as function arguments. However, this is the reason #[repr(transparent)]
exists, and it seems like tempting fate not to use it.