Hide `detach_current_thread` method from docs:
This method is hidden because it is almost never needed and its use
requires some extra care. Its status might be reconsidered if we
learn of any use cases that require it.

Also, improve docs on threads attachment/detachment,
clarifying the attachment status on VM instantiation
and the use-cases for explicit detachments and
subsequent effects of operations on invalidated JNIEnvs.

Co-authored-by: Dmitry Timofeev <dmytro.tymofeev@xdev.re>