Call into libjli.dylib instead of libjvm.dylib for MacOS
This is a workaround for the legacy Java SE6 popup on MacOS.

If you don't have the legacy Java6 installed then the call to
JNI_CreateJavaVM() will result in a popup requesting you to install it
and abort, even if you call the libjvm.dylib that's included as part of
JDK8.

According to jpype-project/jpype#160 and
https://bugs.openjdk.java.net/browse/JDK-7131356 this is a Java8 bug and
the recommended fix is to call into libjli.dylib instead.