Restructure, reduce and edit
Edit each exercise as 1. exercise 2. hint 3. solution, which hopefully
gives the reader an opportunity to try it themselves, and to come to
their own conclusions.

Remove the autolocal discussion for now. It will come later.

Introduce arrays and strings earlier, with a practical
exercise (rendering a chart).

Fix a few gradle build settings (parameter passing to tests, and
disabling the test cache).

Simplify the error handling section, and add a wrapper function for the
reader to use.

Rely much more heavily on jni-rs docs and the jni spec.

Add discussion of ManuallyDrop, and different approaches for
implementing the counter.

Add Future Topics section.

Move debugging to the Appendix; many people don't care or use debuggers,
that work with JNI. So it's a useful exercise, but not central to the
book, and we don't make use of it ourselves yet.

Remove the cleaning section - this is already a lot of material. We'll
add it later.

Add discussion of use-after-free, double-free, and never-free to the
Counter section.

Remove playpens for jni methods, where it's not going to work.