Major edits
First, thanks for your patience on this commit history. Lets talk about
whats going on here.

This diff is big because it's introducing some testing. There were some
syntactic errors in the code before, because everything was inline to
mdbook. Now, the starter project has been cloned and worked through in
the completed/ directory. This project contains anchor tags, which the
project refers to. Since we build examples over time, the various
changes sit behind feature flags. A script in the next commit is used
for testing each feature flag in succession.

The new ideas are introduced in the Counter discussion section.
Previously, we just asked a question and then didn't say anything. Now,
we have a details tag that the reader can click, so they can see an
opinion.

There are also some minor improvements to the error handling section.
Previously, we did not provide an example of catch_unwind, we just said
"hey, by the way, you can catch panics." So, try_java has a catch_unwind
now. It also relies on expect instead of unwrap, since occasionally
people will misspell the exception name and get a panic.

A bug was fixed on checkAvailility - previously, it cast Errors to
RuntimeException.

ChartRenderer was changed to contain a static method, since the
constructor was distracting.

Next, I thought it this book was already pretty big, so I removed the
Counter section for now. We're going to tackle that in a later PR. For
now, we just have the exercise description included, which we can adjust
as well.

There were some stylistic changes - Exception is a specific class in
java. To reduce confusion with checked Exception, the book uses
exception (lowercase) instead.

In the introduction, we added a suggestion to the reader that they use
cargo watch in combination with gradle so that they have a better
development workflow. This makes a big quality of life difference,
especially for readers that don't use Java IDEs.

There are various editing improvements. I edited the linking section
down quite a bit more. I think it flows a lot better than it did
previously. The debugging section in the appendix got a few minor
editing improvements.