JFieldID: remove lifetime and impl Send + Sync
Field IDs by their nature are intended to be cached by applications
which was generally precluded by the lifetime parameter. According
to the JNI spec, field IDs are also not garbage collected as previously
documented and they remain valid until their corresponding class is
unloaded.

Since the real lifetime details can't be represented via a Rust lifetime
(and it would be an excessive cost at this level of abstraction to pair
every field ID with a global reference to the class) then this patch
removes the Rust lifetime and adds "Safety" documentation to explain
what developers need to do if caching field IDs.

It's perhaps also worth noting that the very conservative lifetime
parameter that existed before wouldn't have provided any technical
safety against a method's class being unloaded within that lifetime.

As field IDs are not tied to a thread / JNIEnv it also makes sense to
declare them as Send + Sync safe.

Fixes: #363