benches: Test LocalDateTime::of (7 args + returning a object)
This adds some additional benchmarks that call one of the `of`
constructors for a `LocalDateTime` that takes 7 arguments.

This makes sure we test something that returns an object, considering
that JType::Object() currently requires an owned String which
is expected to be a noticeable overhead.

By testing a method with a decent number of arguments we may also
better highlight the current overhead of collecting arguments into
a Vec when mapping args from JValue to jvalue.