Mark all rust_field APIs as `unsafe`
The APIs `JNIEnv::set_rust_field`, `JNIEnv::get_rust_field` and
`JNIEnv::take_rust_field` are all _very_ unsafe and rely on the caller
being sure that the field will contain a valid pointer for `T` and that
the field isn't somehow copied by Java and then taken multiple times by
Rust.

In addition to marking each function as `unsafe` the documentation has
been updated to add a `# Safety` section for each function.

The documentation has also been updated to take into account issue #336
by suggesting the use of `Closeable` instead of any use of a `finalizer`
for cleaning up Rust resources.

The documentation also tries to clarify that the Java object is locked
whenever a rust field is set or accessed.

Instead of implying that leaking memory is unsafe (as the previous docs
did) they now just highlight the fact that `set_rust_field` may lead
to a leak if `take_rust_field` is not called in the future.

Fixes: #219
Fixes: #336