Only keep the invocation-dyn code
This removes the build-time linking against a JVM via the 'invocation'
feature in favour of runtime / dynamic linking (aka 'invocation-dyn')

The 'invocation-dyn' feature name was dropped, so we go back to just
having a single 'invocation' feature.

This also updates the documentation in line with this change and adds
a Changelog entry.