Remove the static global JVM cache
This global cache could potentially interfere with applications calling
`with_libjvm()` more than once with different paths if the path will
simply be ignored for subsequent calls.

We can hopefully trust that the OS will handle caching as necessary so
we we don't need to worry about the redundant cost of reloading the same
library.