Implement Deref<Target=[T]> for AutoArray and AutoPrimitiveArray
This makes it possible to read and write the elements of a Java array
without resorting to unsafe code.

Both AutoArray and AutoPrimitiveArray now need to be constructed with
a known array length (::new() will query the length and an unsafe
::new_with_len() API could be used in case the length was already known)

The `.size()` API has been replace with `.len()` and `.is_empty()` and
`.len()` doesn't return a Result since it's no longer possible to
create an `Auto[Primitive]Array` without a known length.