Adds JavaVM::destroy() API
This adds an (unsafe) `JavaVM::destroy()` API that can be used to try
and unload the JavaVM on supported platforms.

This adds a `dev-dependency` on the `rusty-fork` crate to make it easy
to test `JavaVM::destroy()` in a separate process, which is needed to
avoid breaking other tests.

It's notable that the way in which `jni-rs` uses an
`InternalAttachGuard` to automatically detach threads makes it
particularly difficult to use `JavaVM::destroy()` safely due to the fact
that daemon threads will automatically try and access an invalid
`JavaVM` pointer if they exit after `JavaVM::destroy()` has been called.

This also exposes `JavaVM::detach_current_thread` (it was previously
public but hidden from docs) since that provides a mechanism for daemon
threads to potentially detach themselves from a `JavaVM` explicitly
if they need to be able to continue running after `JavaVM::destroy()`
has been called.

Closes: #391