Removes deprecation attribute for GlobalRef::as_obj
This was recently marked as deprecated since landing #398
meant we could also get a reference to a JObject via an implementation
of `AsRef<JObject>`.

Since there are multiple `AsRef<T>` implementations for `GlobalRef`
there are times when the compiler isn't able to infer which
implementation of `.as_ref()` you are trying to call, and so it's
still convenient to have a less ambiguous `.as_obj()` API that can
be used instead.

Closes: #408