Adds call_new_object_with_array_class() test
This test verifies that we get a safe, runtime error if we try and
instantiate an array object via `env.new_object()` considering that
the JNI spec states that `NewObjectA` "must not refer to an array
class".

Our expectation is that `env.new_object` will never get as far as
calling `NewObjectA` in this case since it will first fail to
find any constructor `MethodId` for an array class; consistent
with how Java's `Class::getConstructors()` method doesn't expose
array class constructors.