Make get_object_class take a mutable reference
Since this API creates a new local reference it needs to take a mutable
reference to the JNIEnv to be sure it's safe to create a new
local reference with the `'local` lifetime associated with that
`JNIEnv`.

Without a mutable reference then it's possible that the `JNIEnv` isn't
associated with the top JNI stack frame.