Ports jni to the proposed jni-sys 0.4 API
The two notable changes in jni-sys 0.4 are that JNINativeInterface
is now a union that namespaces the functions by the versions
they are valid for like:

```rust
struct JNINativeInterface__1_1 {
    GetVersion:
}
struct JNINativeInterface__1_2 {}
struct JNINativeInterface__1_3 {}
struct JNINativeInterface__reserved {}
union JNINativeInterface {
    v1_1: JNINativeInterface__1_1,
    v1_2: JNINativeInterface__1_2,
    v1_3: JNINativeInterface__1_3,
    reserved: JNINativeInterface__reserved,
}
```

and `jboolean` is now an alias for `bool` instead of `u8`

Since this change is largely about improving how we handle JNI version
checks it made sense to rework the `JNIVersion` type to use more
conventional JNI version designations (so `V1`, `V2`, `V4`, `V6` and
`V8` are now `V1_1`, `V1_2`, `V1_4`, `V1_6` and `V1_8`) and the type is
now a transparent `i32` wrapper that just provides a more ergonomic way
to check a raw JNI version. There is no 'invalid' version any more so
even if `JNIVersion` has not been updated to know about some future
`V30` it is still possible to check the `.major()` and `.minor()`
component and it is valid to do `Ord`ered comparisons between versions.

`JNIEnv::get_version()` was renamed to `JNIEnv::version()` since that's
more idiomatic for a Rust API.

This was used as an opportunity to quite significantly rework the
low-level macros used to make jni-sys function calls to make them much
simpler and also safer to use.

Key changes:
- The JNI API version needs to be specified along with every function
  call and will fail to compile if the function doesn't exist for the
  given version.
- No hidden use of `return` in the macros that can lead to very
  surprising / non-obvious control flow. Macros evaluate to a single
  value that may be a `Result` such that we can use with `?` outside of
  the macro for early returns.
- The `catch` macro is removed since we don't have to worry about
  how some macros might use `return` in confusing ways.
- No hidden use of `unsafe` around the most dangerous code! These
  macros are all very unsafe to use and all callers need to have
  their own `unsafe` blocks and should ideally document how
  safety has be considered.
- An infallible JNI function, like `GetVersion`, can now be called
  without getting a redundant `Result` to unwrap.
- The macros take `JNIEnv` and `JavaVM` arguments instead of raw
  pointers which means we don't need `null` pointer checks for
  each function call.
- Adds more `// Safety:` comments
- Added missing null checks for a number of function calls
  (e.g. `is_assignable_from()` and `is_instance_of()`)

The version namespacing has highlighed several places where the jni
crate was not checking the version appropriately.

This now makes it so JNI 1.4 is the minimum supported version of JNI
since we couldn't even call ExceptionCheck without JNI 1.2 and
requiring >= 1.4 means we don't need any runtime version checks for
the direct byte buffer APIs.

The following functions are now infallible:
  - `JNIEnv::version`
  - `JNIEnv::exception_check`
  - `JNIEnv::exception_clear`
  - `JNIEnv::exception_describe`
  - `JNIEnv::is_same_object`
  - `JNIEnv::delete_local_ref`
  - `WeakRef::is_same_object`
  - `WeakRef::is_weak_ref_to_same_object`
  - `WeakRef::is_garbage_collected`

Fixes: #461
Fixes: #463