Port from windows-sys to windows-bindgen (#552)
This implements an alternative to #513 that requires less maintenance overall
and has a lower chance of bitrotting in various downstream dependency
trees/lock files.

In summary, the changes proposed here move away from using the pre-generated
`windows-sys` crate to using the `windows-bindgen` tool to make the minimal Windows
binding code required for jni, which only uses a very small surface area.

This introduces a small, manually run CLI, (`windows-bindings`) which uses
`windows-bindgen` to create the required bindings that mimic the `windows_sys`
bindings.

This also vendors a little bit of raw-dylib code from `windows-link` in the form
of a `win_link` macro (which is not yet released) because we still need to slot
into the expected structure that `windows-bindgen` looks for.

In the future we could avoid copying the `win_link` macro and add a dependency
on the `windows-link` crate because that crate has no dependencies and should
never have its minor version bumped.

Closes: #513, #563, #565