In `JavaVM::{new,with_libjvm}`, prevent libjvm from being unloaded. (#554)
If libjvm is unloaded while it's running, the process will crash as soon as the
JVM tries to do anything, since all of its code has been unloaded.

`JavaVM::{new,with_libjvm}` unwittingly unloaded the libjvm, by dropping the
`libloading::Library` that was used to load it.

This uses `std::mem::forget` on the `libloading::Library` to prevents libjvm
from ever being unloaded.