Add JavaVM::singleton() getter
Since merging #531 we have committed to assuming that the JVM is a
singleton that can never be unloaded/reloaded and these updated
implementations.

Currently there are various places where we save temporary vm pointers
(such as GlobalRef) that shouldn't be necessary if we can instead assume
a VM must have already been initialized and its pointer is available
globally.

This adds a `JavaVM::singleton()` API that can `clone()` a lazily
initialized `JavaVM` stored in a static variable.

Currently `JavaVM::singleton()` is initialized when `JavaVM::from_raw` is
called (which is used internally whenever we create a `JavaVM` wrapper).

As follows ups, the plan is:

- To ensure that `JNIEnv` depends on `JavaVM::singleton` being initialized
  (so it's easy for code to depend on `JavaVM::singleton` once they have
  observed a `JNIEnv` pointer)
- Update APIs like `GlobalRef` and `WeakRef` so they don't need to save an
  internal `JavaVM` pointer for their `Drop` implementations.

For now, code that has access to JNIEnv can ensure that JavaVM::singleton()
is initialized by doing:

```rust
let _vm = env.get_java_vm();
```