Add JavaVM::singleton() getter
Since merging #531 we have committed to assuming that the JVM is a
singleton that can never be unloaded/reloaded and these updated
implementations.

Currently there are various places where we save temporary vm pointers
(such as GlobalRef) that shouldn't be necessary if we can instead assume
a VM must have already been initialized and its pointer is available
globally.

This adds a `JavaVM::singleton()` API that can `clone()` a lazily
initialized `JavaVM` stored in a static variable.

Although we can't always just assume `JavaVM::singleton()` is safe to
call - if we have a type that, for example, requires a JNIEnv to
construct then we can use that to infer that `JavaVM::singleton()` must
be initialized and would then be accessible for all type methods.

As a follow up; the plan is to use this to remove the need for saving
a vm pointer within each GlobalRef.