Remove DerefMut implementation for AutoLocal
I can't think of a good reason for needing to be able to mutate the
inner reference of an AutoLocal.

It could let you swap the reference that will be automatically deleted
but to me that seems like that would be confusing to read compared to
explicitly using unwrap/forget on one reference and using `.auto()` to
wrap the other reference.

Previously AutoLocal contained a JNIEnv pointer so maybe the benefit
of DerefMut was that you could swap a reference in and keep the existing
JNIEnv pointer on the stack.