stash: JavaVM::singleton() returning Result
This has nicer ergonomics considering almost all callers can simply propagate
any error with `?`.

This also specifically avoids a panic in AutoLocal::drop() in case
JavaVM::singleton() returns an error since there is actually a corner
case where this could happen within a native method. In this case it's
fine to skip trying to delete the reference and probably for the best to
avoid a panic in case the native method doesn't properly catch panics.