Have AutoElements own it's (AsRef<JPrimitiveArray>) array reference
This simplifies AutoElements so it is now only parameterized by one
`'array_local` lifetime that corresponds to the JNI stack frame that
owns the local reference for the array whose elements are being
accessed. (or it could be `'static` if given a `GlobalRef`).

Since `AutoElements` needs to keep hold of the array reference so it can
call `Release*ArrayElements` it now does that by taking ownership of a
generic `TArrayRef: impl AsRef<JPrimitiveArray<T>>` type.

Co-authored-by: Jordan Rose <jrose@signal.org>