Add (unsafe) attach_current_thread_guard + expose AttachConfig
This exposes a lower-level (but unsafe) attach_current_thread_guard
API that is used by attach_current_thread and
attach_current_thread_for_scope.

This provides a lower-level escape hatch that gives full control over
the thread attachment but is unsafe because it also directly exposes the
`AttachGuard` that the caller must manage safely.

`attach_current_thread_guard` accepts a (lazily evaluated) closure that
can return an `AttachConfig` that effectively exposes
`JavaVMAttachArgs`.

This enables control over the `Thread` name in the JVM as well as the
`ThreadGroup` that the thread is associated with.

`AttachConfig` can also configure whether the attachment is `scoped` or
not, which controls whether we create a permanent TLS attachment or an
`AttachGuard::from_owned` attachment that will detach when dropped.

`attach_current_thread_guard` is now used in the implementation of both
`attach_current_thread` and `attach_current_thread_for_scope`

Note: the plan is to also add a safe wrapper API that exposes the
`AttachConfig` but not the `AttachGuard`.

Fixes: #268
Fixes: #543