Generalize class + methodID caching + enable safe downcasting
This adds a safe way to downcast local references via
`JNIEnv::downcast_local<T: JObjectRef>()`.

This adds a `lookup_class` method to `JObjectRef` as well as a `const
CLASS_NAME` introduces a `LoaderContext` for influencing how a class is
loaded.

`lookup_class()` is expected to return a borrowed `GlobalRef<JClass>` and
this introduces a new convention for how different reference types cache
a singleton with a global JClass reference along with MethodIDs.

The downcast_local implementation works by calling
JObjectRef::lookup_class for the target, downcast type and then uses
IsInstanceOf to check the given object reference being downcast.

Actual class lookups are handled by `LoaderContext::load_class` which
will first prefer to find a `ClassLoader` and call `Class.forName(name, false, loader)`
before falling back to `FindClass`

If the `LoaderContext` has a direct reference to a `ClassLoader` then
that's the only thing checked.

Otherwise, `load_class()` will first attempt to find a loader with
`Thread.currentThread().getContextClassLoader()` and otherwise if the
`LoaderContext` has a `FromObject` reference it will look up that
object's associated `ClassLoader.

The implementation of `downcast_local` will provide a reference to the
object being downcast as `LoaderContext::FromObject` so that
`load_class` should always be able to find a suitable loader for an
application class (if the downcast is valid).

Fixes: #515