Make JList a JObject wrapper, implementing JObjectRef
JList is now a `#[transparent]` `JObject` wrapper, like `JClass` or
`JString` that implements `JObjectRef`

Instead of looking up `JMethodID`s when creating each `JList`, these are
now cached via a `JListAPI` singleton, consistent with how other
`JObjectRef` types cache their `JClass` and method IDs.

Most of the `JList` API internally casts the list as a `&JCollection`
and calls the corresponding method.

This also adds bindings for the `isEmpty()` and `clear()` methods.

This deprecates `JList::pop()` since that's not a standard Java API
that can't be implemented based on a single method call.

`JList::iter()` now returns a `JIterator`, so this removes the redundant
`JListIter` type.

Notably there's no longer additional state being maintained outside of the
Java Iterator. Previously the iterator was based on indexing into a list
(which could be an O(N) operation for a linked list) and could
potentially hit IndexOutOfBounds exceptions if the list were modified
while iterating.

This adds some basic tests to tests/jlist.rs