Simplify AutoElementsCritical + add JPrimitiveArray methods
Firstly, this is a follow up to #508 that applies the same simplifications to
`AutoElementsCritical` that were made to `AutoElements`.

This also moves more of the `JPrimitiveArray` implementation details into
the `TypeArray` trait so we don't have an inconsistency in how we access
array elements vs array regions.

All of the `JNIEnv::get_*_array_region` and `JNIEnv::set_*_array_region`
methods are now implemented in terms of `TypeArray::get_array_region`
and `TypeArray::set_array_region` respectively.

`JPrimitiveArray` now has `get/set_array_region` methods and all of the
separate `JNIEnv::get_*_array_region` and `JNIEnv::set_*_array_region`
have been marked as deprecated, with a pointer to the `JPrimitiveArray`
methods.