JavaStr -> MUTF8Chars + add JString::mutf8_chars + ::to_string
To help avoid confusion with the `JString` type and `JNIStr/JNIString`
types this renames `JavaStr` to `MUTF8Chars` and the documentation
focuses more on it being an RAII guard that borrows the contents of a
`java.lang.String`.

At the same time this reworks `MUTF8Chars` based on the same principles
that were recently used to simplify `AutoElements` and
`AutoElementsCritical` (in #508 and #623) such that it's based on an
`AsRef<JString>` with fewer lifetimes to juggle. (Creating a `MUTF8Chars`
guard doesn't involve creating a new reference and the pointer we
get from `GetStringUTFChar` isn't tied to a JNI stack frame).

`JNIEnv::get_string` has been deprecated in favour of a
`JString::mutf8_chars()` method that is hopefully recognisable for anyone
looking for a `GetStringUTFChars` like API, but also more disoverable
as a `JString` method.

It's also intended to be possible to add a `UTF16Chars` guard later and
`JString::utf16_chars()` method if we want to expose string contents as
utf16, via `GetStringChars`.

For convenience this adds a `JString::to_string` method that handles
calling `JString::mutf8_chars` and then uses `.to_str().to_owned()` to get
an owned Rust string with the contents of a `JString`.

Fixes: #616