make set_static_field safe + mark _unchecked field apis unsafe
This ensures that set_static_field looks up the field ID internally
instead of trusting that a given field ID is valid.

The two get/set_static_field_unchecked APIs are now marked as unsafe
because they rely on the caller ensuring that the field IDs are valid
for the given class and that given types match the field.

This normalizes some naming and documentation details across the various
get/set field APIs and adds more comprehensive unit tests that cover
each of these functions.

Fixes: #505
Fixes: #628