Implement Display for MUTF8Chars and JString
This makes it even easier to get a Rust `String`, from a borrowed
`MUTF8Chars` guard or directly from a `JString`.

The implementation is based on `JavaVM::singleton()` and
`vm.attach_current_thread_scoped()`.

The general expectation is that having a `JString` almost certainly
implies that `JavaVM::singleton()` is initialized and generally expects
that `vm.attach_current_thread_scoped` will probably be a NOOP.

If a `JString` reference is `null` then it will format as "<NULl>

In the (unlikely) case that you capture a `JString` as a native method
argument and try to format it before anything else has initialized
`JavaVM::singleton` then the string will be formatted as "<JNI Not
Initialied>".

In the (unlikely) case of any other JNI error then it will format as
"<JNI Error>".

In these latter two conditions the implementation will also log an
error as they most likely imply some bug or misuse of the API.

The previous `JString::to_string(env)` method has now been renamed to
`try_to_string`.