Renames JNIEnv -> Env and JNIEnvUnowned -> EnvUnowned
One of the important changes made in #570 was that `JNIEnv` is no longer a
transparent `jni_sys::JNIEnv` pointer wrapper.

One of the difficulties with that change though was that we have to consider
that existing code may be using `JNIEnv` to capture raw pointers passed to
native methods and Rust has no way of explicitly stopping a type from being
used unsafely in an FFI interface.

The previous solution was to move the "real" `JNIEnv` to `env::JNIEnv` and make
the original `JNIEnv` a type alias for the FFI-safe `JNIEnvUnowned` that has a
very loud deprecation warning.

We have also added a `Drop` implementation to help trigger a compiler warning
when the type is used in FFI.

Although this solution would certainly mean that everyone updating to 0.22
would see the deprecation message, there was still some risk that after simply
updating some `use` statements to `use jni::env::JNIEnv` then it would still be
possible to accidentally not update native methods to use `JNIEnvUnowned`.

By renaming `JNIEnv` to `Env` then, not only do we remove a redundant `JNI`
prefix, we also further reduce the risk that someone will not update native
methods to use `JNIEnvUnowned`.

This keeps the `JNIEnv` -> `EnvUnowned` type alias plus the loud deprecation
warning so we get FFI safety by default and a clear explanation of what has
changed.

For code that needs to refer to non-ffi `JNIEnv` type they will need to
explicitly refer to the new `Env` type and there's less risk that that
migration will affect native method implementations.

Of course, it's still possible that a blind find/replace could have the
original problem but that's at least more involved than swapping a single `use`
statement and it should hopefully be more likely that someone can do an
interactive find/replace if they know they have a mixture of FFI/non-FFI
`JNIEnv` usages.

Even if someone does a blind find and replace then so long as they choose to
replace `JNIEnv` with `JNIEnvUnowned` then the compiler can complain and help
you migrate any non-FFI usages since `JNIEnvUnowned` has none of the APIs that
would be expected.

Addresses: #633