Rename GlobalRef + WeakRef to Global and Weak
`GlobalRef` and `WeakRef` have recently (#596) been changed so they are
generic over a reference type and so since there is inherently going to
be some churn to update global + weak reference usage when migrating to
0.22 then it won't make much practical difference to also handle a
rename too.

I feel like the use of `Ref` is a little overloaded in this crate and it
doesn't really help clarify anything.

Yes from a Java/JNI point of view these are wrappers for references, but it's
not like we put `Ref` as a suffix for any of the other reference types like
`JObject`, `JClass`, `JString` etc. We also don't have a `Ref` suffix for the
`AutoLocal` wrapper type.

In Rust APIs a `Ref` type is sometimes used to represent a temporary guard that
is borrowing from something which could make these names misleading (e.g. a
`Ref` from a `RefCell` or an `EntryRef` from a `hashbrown::HashMap`).

All things equal, the names "Global" and "Weak" seem like they would be just as
clear and slightly less verbose.

This adds deprecated type aliases to at least help sign post that these have
been renamed when migrating to 0.22

Addresses: #633