Rename AutoLocal -> Auto
Since the introduction of the `IntoAuto` trait, there's a more ergonomic
way to wrap a local reference so that it's automatically deleted when
dropped. For example:

```rust
use jni::objects::IntoAuto as _;
for i in 0..1000 {
    // Ensure we aren't left with a new local for each iteration by
    // wrapping the reference in an `Auto` wrapper.
    let auto_delete_string = env.new_string(c"Hello, world!")?.auto();
}
```

with that in mind (if some level of churn around this part of the API
already exists) it felt like a good opportunity to also move to a more
succinct name.

Even though `Auto` does only apply to local references it doesn't seem
necessary to spell that out in the type name.

The `Global` and `Weak` docs make it clear enough that these will delete
their wrapped reference when dropped, such that it should be clear that
attempting to wrap these in `Auto<T>` would be redundant.

It's nice to have a slightly shorter name in case you're typing out the
full `T` parameter too, like `Auto<JString<'local>>`

This makes a pass over the documentation related to Auto to hopefully
clarify a few details.

This adds a deprecated type alias for `AutoLocal` that can sign post
that the type has been renamed without technically breaking the API yet.

Closes: #633