Have JObjectRef::lookup_class take an `&Env` not `&JavaVM`
The implementations of `JObjectRef::lookup_class` all expect that the
current thread is attached which can only be guaranteed if they get an
`&Env`, not a `&JavaVM`.

Internally the implementations will still end up using `env.get_java_vm()`
followed by `vm.with_env_current_frame()` so that `lookup_class` doesn't
need a mutable Env reference, but it still makes sense to be passed an
Env reference to prove that the current thread is attached.

Fixes: #639