Have `AutoElements[Critical]::discard()` take ownership and drop
It seems logical that if you've explicitly decided to "discard" the
elements you have then you're probably ready for the pointer to be
released and for the `AutoElements` to be dropped.

This updates `discard()` to take ownership of `self` so it can be
dropped after setting the mode to `NoCopyBack`.

Fixes: #402