Check `Env::thow_new` class and add `Env::throw_new_void`
This adds an `IsAssignableFrom` check in `Env::throw_new` to make sure
that the given class is a `java.lang.Throwable` subclass.

The JNI spec says that the class must be a Throwable and does not define
can happen if it isn't.

On the other hand, the JNI spec does allow for the message argument to be null
and although it's not well specified, that will result in the exception
being constructed with no arguments.

To expose that functionality, this adds an `Env::thow_new_void` that
only takes a `class`, without a `msg` argument.

Alternatively we could wrap the `msg` for `throw_new` in an `Option` but
since `msg` is generic it becomes awkward to pass None without
specifying what the `Some` value would have been.

Also, since it's probably much more common to throw with a message, it
seems worthwhile not making that common case more verbose by needing to
wrap every message in `Some()`.

Fixes: #644
Fixes: #486