Add `Env::as_cast_raw` for casting raw JNI references
This makes it possible to borrow a raw JNI jobject pointer and cast it
as a type-safe `jni` reference with a runtime `IsInstanceOf` check.

This also makes it possible to borrow raw global references without
acquiring any kind of owned reference type that could accidentally be
used to deleted the reference.

Fixes: #658