Adds `EnvOutcome` with error handling via `ErrorPolicy` trait
Instead of returning a Result from `EnvUnowned::with_env` we now return
an `EnvOutcome` which pairs an `Outcome` with a raw `JNIEnv` pointer.

An `Outcome` is like a `Result` except it also has a `Panic` variant.

An `EnvOutcome` can be resolved from an `Outcome::Ok(T)/Err/Panic` into
a `T` value that is required for returning from a native method.

This "resolve" is done using a given `ErrorPolicy` trait implementation
so that it's possible to consistently customise how errors and panics
are handled across all of your native methods.

The interesting bits of `ErrorPolicy` are:

```rust
pub trait ErrorPolicy<T, E> {
    /// Runs for any [`Outcome::Err`].
    ///
    /// This must map the error to some value that can be returned by the native
    /// method and may be used to throw a Java exception or log the error.
    ///
    /// If this returns an `Err` then `on_internal_jni_error` will be called.
    ///
    /// If this panics then `on_panic` will be called.
    fn on_error<'unowned_env_local: 'native_method, 'native_method>(
        env: &mut Env<'unowned_env_local>,
        cap: &mut Self::Captures<'unowned_env_local, 'native_method>,
        err: E,
    ) -> crate::errors::Result<T>;

    /// Runs for any [`Outcome::Panic`].
    ///
    /// This must return some value that can be returned by the native
    /// method and can be used to throw a Java exception or log the panic
    ///
    /// If this returns an `Err` then `on_internal_jni_error` will be called.
    ///
    /// If this panics then `on_internal_panic` will be called.
    fn on_panic<'unowned_env_local: 'native_method, 'native_method>(
        env: &mut Env<'unowned_env_local>,
        captures: &mut Self::Captures<'unowned_env_local, 'native_method>,
        payload: Box<dyn std::any::Any + Send + 'static>,
    ) -> crate::errors::Result<T>;
}
```

This provides a reusable `ThrowRuntimeExAndDefault` implementation that will
throw any error or panic as a `java.lang.RuntimeException` and return a
default value.

So then a native method may be implemented something like:

```rust
#[no_mangle]
pub extern "system" fn Java_HelloWorld_hello<'caller_local>(
    mut unowned_env: EnvUnowned<'caller_local>
) -> JObject<'caller_local> {
    unowned_env.with_env(|env| -> jni::errors::Result<JObject> {
        // do stuff that might fail or panic
        Ok(JObject::null()) // placeholder
    }).resolve::<ThrowRuntimeExAndDefault>()
}
```

This change also removes the `Error::PanicCaught` variant since we no
longer automatically convert panics into jni Errors.

## AttachGuard simplification

These changes led me to update `AttachGuard` so that it's now just a
thin wrapper around the `Env` type and it has its own lifetime
that is passed through to the inner `Env`.

This change simplifies being able to route the lifetime that may come
from `EnvUnowned` through to the upgraded `Env` without having to deal
with the unsafe lifetime transmute that existed before.

From a developer's pov almost nothing changes, unless you're using one
of the `unsafe` APIs to create an `AttachGuard`. The rules haven't
really changed but the caller has some extra responsibility to ensure
they don't associate the guard with an inappropriate lifetime.

There is now an `AttachGuard::borrow_env_mut()` API for borrowing the inner
`&mut Env` that asserts that the AttachGuard represents the top of the
stack.

The various AttachGuard::with_env wrappers have been removed since they
were just trivial wrappers that added no safety or functionality.

## Documentation

This tries to update some of the relevant AttachGuard and Env and lib.rs
documentation and examples and also includes a few fly-by documentation
changes made at the same time.

Fixes: #660
Addresses: #432
Addresses: #485